pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["Ricardo Dani <ricardodani@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {version = "^0.58.1", extras = ["all"]}
tortoise-orm = "^0.16.13"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
pytest = "^5.4.3"
pylint = "^2.5.3"
pytest-stub = "^1.1.0"
asynctest = "^0.13.0"
flake8 = "^3.8.3"
aiofiles = {git = "https://github.com/Tinche/aiofiles/", rev = "master"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
